+++ searchengine-20211025/core/Models/ListLogsRequest.cs
@@ -1,0 +1,38 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Searchengine20211025.Models
+{
+    public class ListLogsRequest : TeaModel {
+        [NameInMap("endTime")]
+        [Validation(Required=false)]
+        public string EndTime { get; set; }
+
+        [NameInMap("pageNum")]
+        [Validation(Required=false)]
+        public string PageNum { get; set; }
+
+        [NameInMap("pageSize")]
+        [Validation(Required=false)]
+        public string PageSize { get; set; }
+
+        [NameInMap("query")]
+        [Validation(Required=false)]
+        public string Query { get; set; }
+
+        [NameInMap("startTime")]
+        [Validation(Required=false)]
+        public string StartTime { get; set; }
+
+        [NameInMap("type")]
+        [Validation(Required=false)]
+        public string Type { get; set; }
+
+    }
+
+}